length&parse_ident&ltrim&trim&rtrim
